requirements.txt
firecrawl-py
openai==1.2.3
python-dotenv
pandas
openpyxl